refactor: move metadata and schema types to schema file
