+++ src/lib/external/mud/schema.ts
@@ -1,7 +1,9 @@
 import type { SchemaAbiTypeToPrimitiveType } from "@latticexyz/schema-type/internal";
 import type { Hex } from "viem";
 
+import type { ResourceLabel } from "@/lib/external/mud/common";
 import type { Entity } from "@/lib/external/mud/entity";
+import type { ContractTableDef } from "@/lib/definitions";
 
 /* -------------------------------------------------------------------------- */
 /*                                   SCHEMAS                                  */
@@ -49,6 +51,58 @@ export type UnparsedAbiPropertiesSchema = {
   readonly [k: string]: {
     readonly type: SchemaAbiType;
   };
+};
+
+/* -------------------------------------------------------------------------- */
+/*                                   TABLES                                   */
+/* -------------------------------------------------------------------------- */
+
+export type BaseTableMetadata<M extends Metadata = Metadata> = M & {
+  name: string;
+  globalName: string;
+  namespace?: string;
+  abiKeySchema: { [name: string]: StaticAbiType }; // local tables are given a default key schema as well for encoding/decoding entities
+};
+
+export type ContractTablePropertiesSchema<tableDef extends ContractTableDef> = {
+  __staticData: Type.OptionalHex;
+  __encodedLengths: Type.OptionalHex;
+  __dynamicData: Type.OptionalHex;
+  __lastSyncedAtBlock: Type.OptionalBigInt;
+} & {
+  [fieldName in keyof tableDef["valueSchema"] & string]: Type &
+    SchemaAbiTypeToRecsType<SchemaAbiType & tableDef["valueSchema"][fieldName]["type"]>;
+};
+
+export type ContractTableKeySchema<tableDef extends ContractTableDef> = {
+  [fieldName in keyof tableDef["keySchema"] & string]: Type &
+    SchemaAbiTypeToRecsType<SchemaAbiType & tableDef["keySchema"][fieldName]["type"]>;
+};
+
+export type ContractTableMetadata<tableDef extends ContractTableDef> = {
+  name: tableDef["name"];
+  namespace: tableDef["namespace"];
+  globalName: ResourceLabel;
+  abiPropertiesSchema: { [name in keyof tableDef["valueSchema"] & string]: tableDef["valueSchema"][name]["type"] };
+  abiKeySchema: { [name in keyof tableDef["keySchema"] & string]: tableDef["keySchema"][name]["type"] };
+};
+
+// Used to infer the TypeScript types from the RECS types (from schema)
+export type Properties<S extends Schema, T = unknown> = {
+  [key in keyof S]: MappedType<T>[S[key]];
+};
+
+// Used to infer the TypeScript types from the RECS types (from abi)
+export type Keys<TKeySchema extends AbiKeySchema, T = unknown> = {
+  [key in keyof AbiToSchema<TKeySchema>]: MappedType<T>[AbiToSchema<TKeySchema>[key]];
+};
+
+// Used to infer the TypeScript types from the RECS types (excluding metadata)
+export type PropertiesSansMetadata<S extends Schema, T = unknown> = {
+  [key in keyof S as Exclude<
+    key,
+    "__staticData" | "__encodedLengths" | "__dynamicData" | "__lastSyncedAtBlock"
+  >]: MappedType<T>[S[key]];
 };
 
 /* -------------------------------------------------------------------------- */
