+++ src/lib/external/mud/systems.ts
@@ -1,10 +1,10 @@
 import { concat, EMPTY, from, Observable } from "rxjs";
 
-import type { BaseTable, Properties } from "@/tables/types";
+import type { BaseTable } from "@/tables/types";
 import type { TableUpdate, TableWatcherParams } from "@/queries/types";
 import type { Entity } from "@/lib/external/mud/entity";
 import { queries, type QueryFragment } from "@/lib/external/mud/queries";
-import type { Schema } from "@/lib/external/mud/schema";
+import type { Properties, Schema } from "@/lib/external/mud/schema";
 import { tableOperations } from "@/lib/external/mud/tables";
 import type { World } from "@/lib/external/mud/world";
 const { getTableEntities, setEntity, removeEntity, toUpdateStream } = tableOperations;
