+++ src/tables/core/createTable.ts
@@ -1,10 +1,10 @@
 import { Subject } from "rxjs";
 
 import { createTableMethods } from "@/tables/methods/createTableMethods";
-import { type BaseTableMetadata, type Table, type BaseTable } from "@/tables/types";
+import { type Table, type BaseTable } from "@/tables/types";
 import { getEntityHex, type EntitySymbol } from "@/lib/external/mud/entity";
 import { createIndexer } from "@/lib/external/mud/indexer";
-import type { Schema } from "@/lib/external/mud/schema";
+import type { BaseTableMetadata, Schema } from "@/lib/external/mud/schema";
 import { mapObject, transformIterator } from "@/lib/external/mud/utils";
 import type { World } from "@/lib/external/mud/world";
 import { uuid } from "@/lib/external/uuid";
