+++ src/tables/methods/createTableMethods.ts
@@ -3,10 +3,10 @@ import { useEffect, useState } from "react";
 
 import { createTableKeyMethods } from "@/tables/methods/createTableKeyMethods";
 import { createTableWatcher } from "@/tables/methods/createTableWatcher";
-import type { BaseTable, BaseTableMetadata, Properties, PropertiesSansMetadata, TableMethods } from "@/tables/types";
+import type { BaseTable, TableMethods } from "@/tables/types";
 import { type TableWatcherParams, type TableUpdate, TableMethodsWatcherOptions } from "@/queries/types";
 import { defaultEntity, type Entity } from "@/lib/external/mud/entity";
-import type { Schema } from "@/lib/external/mud/schema";
+import type { BaseTableMetadata, Properties, PropertiesSansMetadata, Schema } from "@/lib/external/mud/schema";
 import { type TableMutationOptions, tableOperations } from "@/lib/external/mud/tables";
 import { queries } from "@/lib/external/mud/queries";
 import type { World } from "@/lib/external/mud/world";
